% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishMap-package.R
\docType{package}
\name{FishMap-package}
\alias{FishMap}
\alias{FishMap-package}
\title{FishMap: Spatial-temporal Models For Fish Distributions with 'TMB'}
\description{
Combine survey and VMS data to model the distribution of fish using 'TMB' models.
}
\author{
\strong{Maintainer}: Baptiste Alglave \email{balglave@uw.edu} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

Authors:
\itemize{
  \item Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})
  \item Swann Floc'hlay \email{swann@thinkr.fr} (\href{https://orcid.org/0000-0003-1477-830X}{ORCID})
}

Other contributors:
\itemize{
  \item Ifremer [copyright holder]
}

}
\keyword{internal}
