% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_generate_graphs.R
\name{fm_generate_graphs}
\alias{fm_generate_graphs}
\title{generate output graphs}
\usage{
fm_generate_graphs(fm_model_results)
}
\arguments{
\item{fm_model_results}{list Named list obtained from `fm_fit_model()`}
}
\value{
list A named list of all generated graphs (step 3)
}
\description{
generate output graphs
}
