% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_fit_model.R
\name{fm_fit_model}
\alias{fm_fit_model}
\title{Compile model and fit to data}
\usage{
fm_fit_model(
  fm_data_inputs,
  SE = 1,
  data_source = 1,
  data_obs = 2,
  samp_process = 0,
  b_constraint = 2,
  const_spphab = 1,
  cov_samp_process = 0,
  biomass_temporal = 1,
  sampling_temporal = 0,
  anisotropy = 0,
  lf_link = 0,
  ref_data = "com",
  EM = "est_b",
  month_ref = 1,
  compute_sd = FALSE,
  Version = ""
)
}
\arguments{
\item{fm_data_inputs}{list Named list obtained from `fm_load_data()`}

\item{SE}{integer run ADREPORT - 0: no, 1: yes}

\item{data_source}{integer 1: integrated model (scientific + commercial data), 2: scientific model, 3: commercial model}

\item{data_obs}{integer observation model for biomass - 1: zero_inflated gamma, 2: zero-inflated lognormal, 3: lognormal}

\item{samp_process}{integer Sampling process - 0: no sampling process, 1: inhomogeneous Poisson point process}

\item{b_constraint}{integer put constraint on b parameters - 1: b are positive, 2: no constraints}

\item{const_spphab}{integer Species-habitat relationship - 1: constant in time}

\item{cov_samp_process}{integer covariate in the sampling process - 0: none, 1: covariate in the samplign process}

\item{biomass_temporal}{integer Account for temporal correlation in biomass - 0: no, 1: AR1}

\item{sampling_temporal}{integer Account for temporal correlation in sampling process - 0: no, 1: AR1}

\item{anisotropy}{integer Account for anisotropy}

\item{lf_link}{integer link function of the latent field - 0: log (biomass data), 1: logit (presence absence data)}

\item{ref_data}{character reference data - "com": commercial (default) or "sci": scientific}

\item{EM}{character "est_b": b is estimated, "fix_b": b is fixed}

\item{month_ref}{integer reference month (default is 1 for January)}

\item{compute_sd}{logical compute SD report (time-consumming)}

\item{Version}{version for fm_build_data_params_map}
}
\value{
list A named list of all necessary outputs for generating graphs (step 3)
}
\description{
Compile model and fit to data
}
