% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_generate_graphs.R
\name{fm_generate_graphs}
\alias{fm_generate_graphs}
\title{generate output graphs}
\usage{
fm_generate_graphs(fm_model_results)
}
\arguments{
\item{fm_model_results}{list Named list obtained from `fm_fit_model()`}
}
\value{
list A named list of all generated graphs (step 3)
}
\description{
generate output graphs
}
\examples{
\donttest{

# loading `fm_fit_model()` outputs
fm_fit_model_outputs <- readr::read_rds(
  system.file("examples", "part2_output_small.rds",
              package = "FishMap")
)

# run function
fm_generate_graphs(fm_model_results = fm_fit_model_outputs)

}
}
