% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_load_data.R
\name{fm_load_data}
\alias{fm_load_data}
\title{Load and prepare data for model fitting}
\usage{
fm_load_data(
  species,
  fleet,
  fitted_data = c("biomass", "presabs"),
  survey_data_file,
  vmslogbook_data_file,
  study_domain_file,
  year_start,
  year_end,
  month_start,
  month_end,
  time_step = c("Month", "Quarter"),
  k,
  grid_xmin,
  grid_xmax,
  grid_ymin,
  grid_ymax,
  seed = 29510
)
}
\arguments{
\item{species}{character Species of interest}

\item{fleet}{character Fleet chosen according to the species of interest. A fleet is considered to have homogeneous catchability and targeting behavior.}

\item{fitted_data}{character Type of the data to be fitted to the model (either `biomass` for biomass data - positive-continuous data - or `presabs` for presence-absence data). Default is `biomass`}

\item{survey_data_file}{character File containing the scientific (survey) data}

\item{vmslogbook_data_file}{character File containing the commercial (vmslogbook) data}

\item{study_domain_file}{character File containing the data describing the study area}

\item{year_start}{integer Starting year}

\item{year_end}{integer Ending year}

\item{month_start}{integer Starting month}

\item{month_end}{integer Ending month}

\item{time_step}{character Time step for the model (either `Month` for monthly time step or `Quarter` for quarterly time step). Default is `Month`}

\item{k}{numeric Parameter controlling the number of knots of the mesh. The higher the denser.}

\item{grid_xmin, grid_xmax, grid_ymin, grid_ymax}{numeric Limitation of the grid for the spatial domain}

\item{seed}{integer The seed controlling for random effect. Default is 29510.}
}
\value{
list A named list of all necessary outputs for model fitting (step 2)
}
\description{
Load and prepare data for model fitting
}
\examples{
# run part1
survey_data_file <- system.file("original_data",
                                "Solea_solea",
                                "survey_data.Rdata",
                                package = "FishMap"
                                )

vmslogbook_data_file <- system.file("original_data",
                                "Solea_solea",
                                "vmslogbook_data.Rdata",
                                package = "FishMap"
                                )

study_domain_file <- system.file("original_data",
                                "Solea_solea",
                                "study_domain.Rdata",
                                package = "FishMap"
                                )

fm_data_inputs <- fm_load_data(species = "Solea_solea",
                         fleet = c("OTB_DEF_>=70_0","OTB_CEP_>=70_0","OTT_DEF_>=70_0"),
                         fitted_data = "biomass",
                         survey_data_file = survey_data_file,
                         vmslogbook_data_file = vmslogbook_data_file,
                         study_domain_file = study_domain_file,
                         year_start = 2018,
                         year_end = 2018,
                         month_start = 11,
                         month_end = 11,
                         time_step = "Month",
                         k = 0.25,
                         grid_xmin = -6,
                         grid_xmax = 0,
                         grid_ymin = 42,
                         grid_ymax = 48)
}
