% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_load_data.R
\name{fm_load_data}
\alias{fm_load_data}
\title{Load and prepare data for model fitting}
\usage{
fm_load_data(
  species = "Solea_solea",
  fleet = c("OTB_DEF_>=70_0", "OTB_CEP_>=70_0", "OTT_DEF_>=70_0"),
  survey_data_file = "survey_data.Rdata",
  vmslogbook_data_file = "vmslogbook_data.Rdata",
  study_domain_file = "study_domain.Rdata",
  year_start = 2018,
  year_end = 2018,
  month_start = 11,
  month_end = 11,
  time_step = "Month"
)
}
\arguments{
\item{species}{character Name of the species of interest, match the data folder name}

\item{fleet}{character Vector of conditional operation to filter fleets, the first one will be taken as reference level in the model}

\item{survey_data_file}{character Name of the Rdata file storing scientific observations}

\item{vmslogbook_data_file}{character Name of the Rdata file storing vms x logbooks observations}

\item{study_domain_file}{character Name of the Rdata file storing study domain informations}

\item{year_start}{integer Lower bound year filter to select data}

\item{year_end}{integer Upper bound year filter to select data}

\item{month_start}{integer Lower bound month filter to select data}

\item{month_end}{integer Upper bound month filter to select data}

\item{time_step}{character Time intervals, must be Month or Quarter}
}
\value{
list A named list of all necessary outputs for model fitting (step 2)
}
\description{
Load and prepare data for model fitting
}
\examples{
# run part1
fm_data_inputs <- fm_load_data()

}
